+++ nautilus_core/model/src/instruments/crypto_future.rs
@@ -16,7 +16,7 @@
 use std::hash::{Hash, Hasher};
 
 use nautilus_core::{
-    correctness::{check_equal_u8, check_positive_i64, check_positive_u64},
+    correctness::{check_equal_u8, check_positive_i64, check_positive_u64, FAILED},
     nanos::UnixNanos,
 };
 use rust_decimal::Decimal;
@@ -54,6 +54,7 @@ pub struct CryptoFuture {
     pub taker_fee: Decimal,
     pub margin_init: Decimal,
     pub margin_maint: Decimal,
+    pub multiplier: Quantity,
     pub lot_size: Quantity,
     pub max_quantity: Option<Quantity>,
     pub min_quantity: Option<Quantity>,
@@ -89,6 +90,7 @@ impl CryptoFuture {
         taker_fee: Decimal,
         margin_init: Decimal,
         margin_maint: Decimal,
+        multiplier: Option<Quantity>,
         lot_size: Option<Quantity>,
         max_quantity: Option<Quantity>,
         min_quantity: Option<Quantity>,
@@ -131,6 +133,7 @@ impl CryptoFuture {
             taker_fee,
             margin_init,
             margin_maint,
+            multiplier: multiplier.unwrap_or(Quantity::from(1)),
             lot_size: lot_size.unwrap_or(Quantity::from(1)),
             max_quantity,
             min_quantity,
@@ -161,6 +164,7 @@ impl CryptoFuture {
         taker_fee: Decimal,
         margin_init: Decimal,
         margin_maint: Decimal,
+        multiplier: Option<Quantity>,
         lot_size: Option<Quantity>,
         max_quantity: Option<Quantity>,
         min_quantity: Option<Quantity>,
@@ -188,6 +192,7 @@ impl CryptoFuture {
             taker_fee,
             margin_init,
             margin_maint,
+            multiplier,
             lot_size,
             max_quantity,
             min_quantity,
@@ -198,7 +203,7 @@ impl CryptoFuture {
             ts_event,
             ts_init,
         )
-        .expect("Failed to create CryptoFuture")
+        .expect(FAILED)
     }
 }