+++ nautilus_core/model/src/instruments/crypto_perpetual.rs
@@ -16,7 +16,7 @@
 use std::hash::{Hash, Hasher};
 
 use nautilus_core::{
-    correctness::{check_equal_u8, check_positive_i64, check_positive_u64},
+    correctness::{check_equal_u8, check_positive_i64, check_positive_u64, FAILED},
     nanos::UnixNanos,
 };
 use rust_decimal::Decimal;
@@ -53,6 +53,7 @@ pub struct CryptoPerpetual {
     pub taker_fee: Decimal,
     pub margin_init: Decimal,
     pub margin_maint: Decimal,
+    pub multiplier: Quantity,
     pub lot_size: Quantity,
     pub max_quantity: Option<Quantity>,
     pub min_quantity: Option<Quantity>,
@@ -86,6 +87,7 @@ impl CryptoPerpetual {
         taker_fee: Decimal,
         margin_init: Decimal,
         margin_maint: Decimal,
+        multiplier: Option<Quantity>,
         lot_size: Option<Quantity>,
         max_quantity: Option<Quantity>,
         min_quantity: Option<Quantity>,
@@ -126,6 +128,7 @@ impl CryptoPerpetual {
             taker_fee,
             margin_init,
             margin_maint,
+            multiplier: multiplier.unwrap_or(Quantity::from(1)),
             lot_size: lot_size.unwrap_or(Quantity::from(1)),
             max_quantity,
             min_quantity,
@@ -155,6 +158,7 @@ impl CryptoPerpetual {
         taker_fee: Decimal,
         margin_init: Decimal,
         margin_maint: Decimal,
+        multiplier: Option<Quantity>,
         lot_size: Option<Quantity>,
         max_quantity: Option<Quantity>,
         min_quantity: Option<Quantity>,
@@ -180,6 +184,7 @@ impl CryptoPerpetual {
             taker_fee,
             margin_init,
             margin_maint,
+            multiplier,
             lot_size,
             max_quantity,
             min_quantity,
@@ -190,7 +195,7 @@ impl CryptoPerpetual {
             ts_event,
             ts_init,
         )
-        .expect("Failed to create CryptoPerpetual instance")
+        .expect(FAILED)
     }
 }